Fix puzzle 13 part 1, add part 2
There was a small mistake in the parsing of nested empty sets,
tripping up the comparisons. This didn't matter for the example,
but it did for the actual input. The comparison code now takes
const references instead of pointers, as does the print code.
The unused parent pointers were removed