Complete day 22 part 2
Complete part 2 of day 22 (folding map into a cube and navigating
it). This code works with both the example input and the actual
input. No hardcoding, in principle this code should work for any
input. Both parts complete in ~15ms. Performance could likely
be improved a bit by using a vector instead of hashmap, but that
can always be done at a later stage.