Day 12 part 2
Replaced naive brute-force implementation with a recursive
approach, with lots of early returns. Memoization is used to make
the runtime acceptable (~3s in debug, ~0.5s in release). Threading
could be used to improve the runtime further, or better rejection
of bad candidates. (Or reducing the amount of clones of data).
Perhaps I will revisit later.