Change day21 solution to polynomial fit
Use a polynomial fit to extrapolate the number of visitable
plots for large numbers of steps, as suggested on the solution
thread. Simplifies code quite a bit.
Polynomial fit only starts to work for the test input for
steps = n*rows+6, with n>=4. For the real input it starts right
away at n = 0. The tests pass much faster now.